requirements.txt
aiohttp==3.6.2
discord.py==1.3.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
discord==1.0.1
numpy==1.18.4
